style (store): purge debugging logs and move the helper functions to util file
